added basic RatingForm set up. Stars were moved here from About. Not Tested.
